ErrorMessage now translate and some test adjustment

+++ P3AddNewFunctionalityDotNetCore/Models/Services/ProductService.cs
@@ -155,9 +155,12 @@ namespace P3AddNewFunctionalityDotNetCore.Models.Services
             // TODO what happens if a product has been added to a cart and has been later removed from the inventory ?
             // delete the product form the cart by using the specific method
             // => the choice is up to the student
-            _cart.RemoveLine(GetProductById(id));
+            if (GetProductById(id) != null)
+            {
+                _cart.RemoveLine(GetProductById(id));
 
-            _productRepository.DeleteProduct(id);
+                _productRepository.DeleteProduct(id);
+            }
         }
     }
 }
